rag-system/requirements.txt
pymongo>=4.5.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
fastapi>=0.104.0
httpx>=0.24.0  # Required for TestClient
pydantic>=2.4.0
flask>=3.0.0
pytest-flask>=1.3.0
mongomock>=4.1.2  # For MongoDB mocking in tests 